Start CRUD Client entity

+++ app-crm-system-server/src/main/java/uz/crm/system/entity/catalogs/Address.java
@@ -27,7 +27,7 @@ public class Address extends AbsEntity {
 
     private String extraAddress;
 
-    private Long Latitude;
+    private Double Latitude;
 
-    private Long Longitude;
+    private Double Longitude;
 }
